Simplify binja backend
The binja backend was previously using the simple backend to find branches that
binja seems to miss. This commit makes each backend independent of the others to
simplify the build system changes for supporting multiple backends.
Unfortunately this means that the tests will not be expected to pass with the
binja backend, but that is a separate issue.