Show output as both a runtime vaddr and an offset into an ELF
This makes analyzing the results for statically-linked ELFs easier since the
runtime vaddr can be directly compared against objdump's output without using
`--file-offsets`. For ET_DYN ELFs showing the output as an offset into the ELF
is typically easier for comparing against objdump, but both are shown in either
case for flexibility and to avoid having to parse the ELFs to check if the
program is ET_EXEC or not in the plugin.

This commit also fixes a bug in ET_DYN ELFs in the case where a loadable
segment's vaddrs don't correspond to offsets into the file (i.e. p_offset !=
p_vaddr for the segment corresponding to a callsite or destination).